Move DeleteMonitorModal into separate file
